add LDAM-Loss and myinvarse

+++ MIL_test.py
@@ -1,17 +1,20 @@
 # -*- coding: utf-8 -*-
+import re
 import torch
 import torchvision
 from torch import nn, optim
 import torch.nn.functional as F
+import torch.distributed
 import numpy as np
 import csv
 import os
-import HoriDataset
-import torch.distributed as dist
+import dataloader_svs
 import torch.multiprocessing as mp
-from torch.nn.parallel import DistributedDataParallel as DDP
-import time
+from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score
+import matplotlib.pyplot as plt
 import sys
+import argparse
+from tqdm import tqdm
 
 #正誤確認関数(正解:ans=1, 不正解:ans=0)
 def eval_ans(y_hat, label):
@@ -27,32 +30,38 @@ def makedir(path):
         os.makedirs(path)
 
 def select_epoch(log_file):
+    if not os.path.exists(log_file):
+        exit()
     train_log = np.loadtxt(log_file, delimiter=',', dtype='str')
-    valid_acc = train_log[1:,3].astype(np.float32)
-    acc_list = []
-    num_gpu = 1
-    total_epoch = valid_acc.shape[0]/num_gpu # GPU数で割る
-    for i in range(int(total_epoch)):
-        tmp = valid_acc[i*num_gpu:(i+1)*num_gpu]
-        if i < 5:
-            acc_list.append(10000)
-        else:
-            acc_list.append(np.sum(tmp))
-    return acc_list.index(min(acc_list))
+    if train_log.shape[0]<7:
+        exit()
+    valid_loss = train_log[1:,3].astype(np.float32)
+    valid_loss[:5]=1000
+    return np.argmin(valid_loss)
 
-import random
-import utils
+def update_test_result(test_dir, test_name, epoch_m):
+    test_fn_list = os.listdir(test_dir)
+    for test_fn in test_fn_list:
+        if test_name in test_fn and not f'epoch-{epoch_m}' in test_fn:
+            os.remove(f'{test_dir}/{test_fn}')
+            print(f'{test_fn} is removed')
+
 
 def test(model, device, test_loader, output_file):
     model.eval() #テストモードに変更
 
-    for (input_tensor, slideID, class_label, pos_list) in test_loader:
+    pred_label = []
+    bar = tqdm(total = len(test_loader))
+    for input_tensor, slideID, class_label, pos_list in test_loader:
+        bar.update(1)
         input_tensor = input_tensor.to(device)
         #class_label = class_label.to(rank, non_blocking=True)
         # MILとバッチ学習のギャップを吸収
         for bag_num in range(input_tensor.shape[0]):
             with torch.no_grad():
                 class_prob, class_hat, A = model(input_tensor[bag_num])
+            
+            pred_label.append([class_label, class_hat])
 
             class_softmax = F.softmax(class_prob, dim=1).squeeze(0)
             class_softmax = class_softmax.tolist() # listに変換
@@ -60,68 +69,79 @@ def test(model, device, test_loader, output_file):
             # bagの分類結果と各パッチのattention_weightを出力
             f = open(output_file, 'a')
             f_writer = csv.writer(f, lineterminator='\n')
-            slideid_tlabel_plabel = [slideID[bag_num], int(class_label[bag_num]), class_hat] + class_softmax # [slideID, 真のラベル, 予測ラベル] + [y_prob[0], y_prob[1], y_prob[2]]
+            slideid_tlabel_plabel = ['', slideID[bag_num], int(class_label[bag_num]), class_hat] + class_softmax # [slideID, 真のラベル, 予測ラベル] + [y_prob[0], y_prob[1], y_prob[2]]
             f_writer.writerow(slideid_tlabel_plabel)
             pos_x = []
             pos_y = []
             for pos in pos_list:
                 pos_x.append(int(pos[0]))
                 pos_y.append(int(pos[1]))
-            f_writer.writerow(pos_x) # 座標書き込み
-            f_writer.writerow(pos_y) # 座標書き込み
+            f_writer.writerow(['pos_x']+pos_x) # 座標書き込み
+            f_writer.writerow(['pos_y']+pos_y) # 座標書き込み
             attention_weights = A.cpu().squeeze(0) # 1次元目削除[1,100] --> [100]
-            attention_weights_list = attention_weights.tolist()
-            att_list = []
-            for att in attention_weights_list:
-                att_list.append(float(att[0]))
-            #att_list = attention_weights.tolist()
-            f_writer.writerow(att_list) # 各instanceのattention_weight書き込み
+            att_list = attention_weights.tolist()
+            f_writer.writerow(['attention']+att_list) # 各instanceのattention_weight書き込み
             f.close()
 
-def test_model(train_slide, test_slide):
+SAVE_PATH = '/Dataset/Kurume_Dataset/yhirono/KurumeMIL'
+
+def test_model(gpu, train_slide, test_slide, name_mode, depth, leaf, mag, classify_mode, loss_mode, constant, augmentation):
 
     ##################実験設定#######################################
-    #train_slide = '123'
-    #valid_slide = '4'
-    #test_slide = '5'
-    mag = '40x' # ('5x' or '10x' or '20x')
-    EPOCHS = 10
-    device = 'cuda:0'
+    device = f'cuda:{gpu}'
+    torch.backends.cudnn.benchmark=True #cudnnベンチマークモード
     ################################################################
+    if classify_mode == 'subtype':
+        dir_name = f'subtype_classify'
+    elif leaf is not None:
+        dir_name = classify_mode
+        if loss_mode != 'normal':
+            dir_name = f'{dir_name}_{loss_mode}'
+        if loss_mode == 'LDAM':
+            dir_name = f'{dir_name}-{constant}'
+        if augmentation:
+            dir_name = f'{dir_name}_aug'
+        dir_name = f'{dir_name}/depth-{depth}_leaf-{leaf}'
+    else:
+        dir_name = classify_mode
+        if loss_mode != 'normal':
+            dir_name = f'{dir_name}_{loss_mode}'
+        if loss_mode == 'LDAM':
+            dir_name = f'{dir_name}-{constant}'
+        if augmentation:
+            dir_name = f'{dir_name}_aug'
+        dir_name = f'{dir_name}/depth-{depth}_leaf-all'
+
     # 訓練用と検証用に症例を分割
     import dataset_kurume as ds
-    train_DLBCL, train_FL, train_RL, test_DLBCL, test_FL, test_RL = ds.slide_split(train_slide, test_slide)
-    train_domain = train_DLBCL + train_FL + train_RL
-    test_domain = test_DLBCL + test_FL + test_RL
-    domain_num = len(train_domain)
-    test_dataset = []
-    for slideID in test_DLBCL:
-        test_dataset.append([slideID, 0])
-    for slideID in test_FL:
-        test_dataset.append([slideID, 1])
-    for slideID in test_RL:
-        test_dataset.append([slideID, 2])
+    if classify_mode == 'leaf' or classify_mode == 'new_tree':
+        _, test_dataset, label_num = ds.load_leaf(train_slide, test_slide, name_mode, depth, leaf, classify_mode)
+    elif classify_mode == 'subtype':
+        _, test_dataset, label_num = ds.load_svs(train_slide, test_slide, name_mode)
 
-    log = f'train_log/log_{mag}_train-{train_slide}.csv'
+    log = f'{SAVE_PATH}/train_log/{dir_name}/log_{mag}_train-{train_slide}.csv'
     epoch_m = select_epoch(log)
-    makedir('test_result')
-    result = f'test_result/test_{mag}_train-{train_slide}.csv'
-    makedir('model_params')
-    model_params = f'./model_params/{mag}_train-{train_slide}_epoch-{epoch_m}.pth'
+    print(f'best epoch is {epoch_m}')
 
+    # resultファイルの作成
+    makedir(f'{SAVE_PATH}/test_result/{dir_name}')
+    result = f'{SAVE_PATH}/test_result/{dir_name}/test_{mag}_train-{train_slide}_epoch-{epoch_m}.csv'
+    update_test_result(f'{SAVE_PATH}/test_result/{dir_name}', f'test_{mag}_train-{train_slide}', epoch_m)
+    if os.path.exists(result):
+        print(f'[{dir_name}/test_{mag}_train-{train_slide}_epoch-{epoch_m}.csv] has been already done')
+        exit()
     f = open(result, 'w')
     f.close()
-
-    torch.backends.cudnn.benchmark=True #cudnnベンチマークモード
 
     # model読み込み
     from model import feature_extractor, class_predictor, MIL
     # 各ブロック宣言
     feature_extractor = feature_extractor()
-    class_predictor = class_predictor()
+    class_predictor = class_predictor(label_num)
     # DAMIL構築
     model = MIL(feature_extractor, class_predictor)
-    model.load_state_dict(torch.load(model_params,map_location='cpu'))
+    model_params = f'{SAVE_PATH}/model_params/{dir_name}/{mag}_train-{train_slide}/{mag}_train-{train_slide}_epoch-{epoch_m}.pth'
+    model.load_state_dict(torch.load(model_params, map_location='cuda'))
     model = model.to(device)
 
     # 前処理
@@ -134,6 +154,7 @@ def test_model(train_slide, test_slide):
         train=False,
         transform=transform,
         dataset=test_dataset,
+        class_count=label_num,
         mag=mag,
         bag_num=50,
         bag_size=100
@@ -143,17 +164,36 @@ def test_model(train_slide, test_slide):
         data_test,
         batch_size=1,
         shuffle=False,
-        pin_memory=False,
-        num_workers=1,
+        pin_memory=True,
+        num_workers=4,
     )
 
     # 学習
     test(model, device, test_loader, result)
 
 if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description='This program is MIL using Kurume univ. data')
+    parser.add_argument('train', help='choose train data split')
+    parser.add_argument('test', help='choose valid data split')
+    parser.add_argument('--depth', default=None, help='choose depth')
+    parser.add_argument('--leaf', default=None, help='choose leafs')
+    parser.add_argument('--mag', default='40x', choices=['5x', '10x', '20x', '40x'], help='choose mag')
+    parser.add_argument('--name', default='Simple', choices=['Full', 'Simple'], help='choose name_mode')
+    parser.add_argument('--gpu', default=1, type=int, help='input gpu num')
+    parser.add_argument('-c', '--classify_mode', default='leaf', choices=['leaf', 'subtype', 'new_tree'], help='leaf->based on tree, simple->based on subtype')
+    parser.add_argument('-l', '--loss_mode', default='normal', choices=['normal','invarse','myinvarse','LDAM'], help='select loss type')
+    parser.add_argument('-C', '--constant', default=None)
+    parser.add_argument('-a', '--augmentation', action='store_true')
+    args = parser.parse_args()
 
-    args = sys.argv
-    train_slide = args[1]
-    test_slide = args[2]
+    if args.classify_mode != 'subtype':
+        if args.depth == None:
+            print(f'mode:{args.classify_mode} needs depth param')
+            exit()
 
-    test_model(train_slide, test_slide)
+    if args.loss_mode == 'LDAM' and args.constant == None:
+        print(f'when loss_mode is LDAM, input Constant param')
+        exit()
+
+    test_model(args.gpu, args.train, args.test, args.name, args.depth, args.leaf,
+            args.mag, args.classify_mode, args.loss_mode, args.constant, args.augmentation)
